move class Runner
